Multiple editors as code cell components
