+++ course-react-ts-portfolio/jbook/src/index.tsx
@@ -1,29 +1,11 @@
-import { useState } from 'react';
 import ReactDOM from 'react-dom';
-
-import CodeEditor from './components/code-editor';
-import Preview from './components/preview';
-import bundle from './bundler';
+import CodeCell from './components/code-cell';
 
 const App = () => {
-  const [input, setInput] = useState('');
-  const [code, setCode] = useState('');
-
-  const onClick = async () => {
-    const output = await bundle(input);
-    setCode(output);
-  };
-
   return (
     <div>
-      <CodeEditor
-        initialValue='const a = 1;'
-        onChange={(value) => setInput(value)}
-      />
-      <div>
-        <button onClick={onClick}>Submit</button>
-      </div>
-      <Preview code={code} />
+      <CodeCell />
+      <CodeCell />
     </div>
   );
 };
